sound/va_vco.cpp: Moved generic oscillator functionality from cem3340 to va_vco.
