+++ src/mame/sequential/prophet5.cpp
@@ -704,14 +704,14 @@ void prophet5_voice_device::device_add_mconfig(machine_config &config)
 	if (m_osc_a_sum_cv)
 		m_osc_a_sum_cv->add_route(0, m_osc_a_freq, 1.0 / RES_K(100));  // R4322 (1%, matched to 0.01% with R4321)
 	VA_SCALE_OFFSET(config, m_osc_a_freq)
-		.add_route(0, m_osc_a, 1.0, cem3340_device::INPUT_FREQ);
+		.add_route(0, m_osc_a, 1.0, cem3340_device::INPUT_FREQ_CTRL);
 
 	// Osc A pulse width control. The PW A master sum CV is mixed with the
 	// polymod signal and fed to the PW CV input of the VCO.
 	if (m_pw_a_sum_cv)
 		m_pw_a_sum_cv->add_route(0, m_osc_a_pw, 1.0 / RES_K(100));  // R4163 (1%)
 	MIXER(config, m_osc_a_pw)
-		.add_route(0, m_osc_a, -RES_K(52.3), cem3340_device::INPUT_PW);  // R4162 (1%)
+		.add_route(0, m_osc_a, -RES_K(52.3), cem3340_device::INPUT_PW_CTRL);  // R4162 (1%)
 
 	// Oscillator A chip. The ramp and pulse outputs are fed, via switches, to
 	// the "+" and "-" inputs of an OTA. The switches control whether each
@@ -736,12 +736,12 @@ void prophet5_voice_device::device_add_mconfig(machine_config &config)
 	if (m_osc_b_sum_cv)
 		m_osc_b_sum_cv->add_route(0, m_osc_b_freq, 1.0 / RES_K(100));  // R4205 (1%, matched to 0.01% with R4206)
 	VA_SCALE_OFFSET(config, m_osc_b_freq)
-		.add_route(0, m_osc_b, 1.0, cem3340_device::INPUT_FREQ);
+		.add_route(0, m_osc_b, 1.0, cem3340_device::INPUT_FREQ_CTRL);
 
 	// Osc B pulse width control. The PW B master sum CV is directly fed to the
 	// PW CV input of the VCO.
 	if (m_pw_b_sum_cv)
-		m_pw_b_sum_cv->add_route(0, m_osc_b, 1.0, cem3340_device::INPUT_PW);
+		m_pw_b_sum_cv->add_route(0, m_osc_b, 1.0, cem3340_device::INPUT_PW_CTRL);
 
 	// Oscillator B chip. The setup is similar to that of osc A, except that the
 	// triangle output is also used, and the outputs are also routed to the
@@ -1465,10 +1465,10 @@ void prophet5_audio_device::device_add_mconfig(machine_config &config)
 	// modulation by the LFO. Many of the route gains, and the triangle scale
 	// and offset are computed in update_lfo_mix().
 	CEM3340(config, m_lfo, CAP_U(0.1), RES_M(2.21))  // U376 - C382 (mylar, 5%), R3138 (1%)
-		.set_pw_cv(VPLUS * RES_VOLTAGE_DIVIDER(RES_K(10), RES_K(2)))  // R3110, R3111, 50% PW.
 		.add_route(cem3340_device::OUTPUT_TRIANGLE, m_lfo_tri_center, 1.0)
 		.add_route(cem3340_device::OUTPUT_RAMP, m_lfo_vca, 1.0)
 		.add_route(cem3340_device::OUTPUT_PULSE, m_lfo_vca, 1.0);
+	m_lfo->set_pw_ctrl(VPLUS * RES_VOLTAGE_DIVIDER(RES_K(10), RES_K(2)));  // R3110, R3111, 50% PW.
 	VA_SCALE_OFFSET(config, m_lfo_tri_center)  // U380B (TL082) and surrounding resistors.
 		.set_scale(0).set_offset(0)
 		.add_route(0, m_lfo_vca, 1.0);
@@ -1868,7 +1868,7 @@ void prophet5_audio_device::cv_w(offs_t cv_index, double cv)
 		case CV_UNISON: m_glide_eg->set_target_v(cv); break;
 
 		case CV_LFO_FREQ:
-			m_lfo->set_freq_cc(VPLUS / RES_K(487) + cv / RES_K(110));  // R3135 (1%), R3136 (1%)
+			m_lfo->set_freq_ctrl(VPLUS / RES_K(487) + cv / RES_K(110));  // R3135 (1%), R3136 (1%)
 			LOGMASKED(LOG_LFO, "LFO frequency: %f\n", m_lfo->freq());
 			break;
 		case CV_WMOD_SRC_MIX:
